Added functions for adding, deleting, updating food. Images are now kept in the API. Fixed image URLs in the main menu. Updated backend

+++ backend/ZurnaciApi/Controllers/FoodController.cs
@@ -1,7 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using ZurnaciApi.Data;
+using ZurnaciApi.Handler;
 using ZurnaciApi.Models;
+
 
 namespace ZurnaciApi.Controllers
 {
@@ -24,7 +26,7 @@ namespace ZurnaciApi.Controllers
         }
 
         [HttpGet("{id:int}")]
-        public async Task<ActionResult<Order>> GetFood([FromRoute] int id)
+        public async Task<ActionResult<Food>> GetFood([FromRoute] int id)
         {
             var food = await _context.Foods.FindAsync(id);
 
@@ -36,14 +38,42 @@ namespace ZurnaciApi.Controllers
             return Ok(food);
         }
 
-        [HttpPost]
-        public async Task<ActionResult<Food>> AddFood([FromBody]  Food food)
+        [HttpGet("uploads/foods/{fileName}")]
+        public IActionResult GetFoodImage(string fileName)
         {
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "Uploads", "Foods", fileName);
+            if (System.IO.File.Exists(filePath))
+            {
+                var fileBytes = System.IO.File.ReadAllBytes(filePath);
+                return File(fileBytes, "image/jpeg"); // Dosya tipi gerektiği gibi ayarlanabilir
+            }
+            return NotFound();
+        }
+
+        [HttpPost]
+
+        public async Task<ActionResult<Food>> AddFood([FromForm] Food food, IFormFile imageFile)
+        {
+            if (imageFile != null)
+            {
+                var imageName = await FoodImageUploadHandler.UploadImageAsync(imageFile);
+                food.Image = imageName;
+            }
+
             _context.Foods.Add(food);
             await _context.SaveChangesAsync();
 
             return CreatedAtAction(nameof(GetFoods), new { id = food.Id }, food);
         }
+        
+        //public async Task<ActionResult<Food>> AddFood([FromBody]  Food food)
+        //{
+        //    _context.Foods.Add(food);
+        //    await _context.SaveChangesAsync();
+
+        //    return CreatedAtAction(nameof(GetFoods), new { id = food.Id }, food);
+        //}
+
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteFood([FromRoute] int id)
         {
@@ -53,11 +83,21 @@ namespace ZurnaciApi.Controllers
                 return NotFound();
             }
 
+            string filePath = Path.Combine(Directory.GetCurrentDirectory(), "Uploads", "Foods", food.Image);
+
+            bool isDeleted = FileDeleteHandler.DeleteFile(filePath);
+
+            if (!isDeleted)
+            {
+                return BadRequest("File could not be deleted.");
+            }
+
             _context.Foods.Remove(food);
             await _context.SaveChangesAsync();
 
             return NoContent();
         }
+
         [HttpPut("{id:int}")]
         public async Task<IActionResult> UpdateFood(int id, [FromBody] Food food)
         {
